api/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
grpcio==1.18.0
grpcio-tools==1.18.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
protobuf==3.6.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
